Move StoreProviderListingInDynamo and type to own pacakge
